requirements.txt
aiofiles==25.1.0
aiosqlite==0.21.0
aiohappyeyeballs==2.6.1
alembic==1.16.5
asyncpg==0.30.0
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
dataclasses-json==0.6.7
Faker==37.0.0
fastapi==0.135.1
frozenlist==1.8.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
iniconfig==2.3.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.13
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.20
langchain-text-splitters==1.1.1
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.21
marshmallow==3.26.2
multidict==6.7.1
mypy_extensions==1.1.0
numpy==2.4.3
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pluggy==1.6.0
propcache==0.4.1
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
reportlab==4.4.2
python-dotenv==1.2.2
PyYAML==6.0.3
redis==7.3.0
requests==2.32.5
requests-toolbelt==1.0.0
SQLAlchemy==2.0.48
starlette==0.52.1
tenacity==9.1.4
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.42.0
websockets==16.0
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0

# Optional benchmark dependencies (install with: `pip install -e .[benchmarks]`)
# deepteam
# garak
# pyrit  # optional; may conflict with strict pins in this repo depending on versions
